requirements.txt
Django==4.2.6
djangorestframework==3.14.0
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.4.0
psycopg2==2.9.7
Pillow==10.0.1
sqlparse==0.4.4
pre-commit
